Remove CovidTestsHistogram component
